requirements.txt
pipecatcloud
pipecat-ai[cartesia,daily,openai,silero,deepgram]
python-dotenv~=1.0.1
